+++ l10n_br_mdfe/models/__init__.py
@@ -1,8 +1,13 @@
+# -*- coding: utf-8 -*-
+# Copyright 2018 KMEE INFORMATICA LTDA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
 from . import sped_operacao
 from . import sped_documento
 from . import l10n_br_mdfe_condutor
 from . import l10n_br_mdfe_lacre
 from . import l10n_br_mdfe_seguro
+from . import l10n_br_mdfe_seguro_averbacao
 from . import l10n_br_mdfe_item
 from . import sped_veiculo
 from . import sped_empresa
